fix: destructure state and remove pages/proposals page
